Fix legacy error: handle lone atoms in scene
Background: In the past, it was possible to copy atoms without their
molecules. This would lead to top-level atoms, which is inconsistent
with the data model invariants (atoms should always be part of a
molecule -- although this requirement might get dropped in the future).
Here, we just wrap a molecule around these atoms for backward
compatibility.