Slightly fix installer arrangement of checkboxes
Note: It still seems that when changing the installer mode
(from install to update or uninstall and back or so) that the components
are not properly updated.
This, however, appears to be an issue with the installer framework itself,
as some labels in the sidebar are also only updated after another change
of the installer mode.